API Routing Fix (#6)
* Startup: Configure routing

I believe something to do with the jump to with the jump to .NET Core
3.1 required us to reconfigure this. This adds the middleware necessary
to route to the API controllers. It also adds a `TestController` as a
reference on how to add API endpoints.

* Remove useless files

We don't need any views since this is just a Web API. The
`DummyController` was just a demo and the corresponding table won't be
in our new database, so we should remove it. The sample data controller
should also be removed since we now have the `TestController` for
reference.

* Project: Remove unnecessary property groups

These are left over from when we had our front-end code in the same
project.

* API routing: Add default route attribute

* APIController attribute: Add to assembly

This gives us some nice perks discussed here:
https://www.strathweb.com/2019/01/enabling-apicontroller-globally-in-asp-net-core-2-2/