Create Skeleton Controllers with Dummy Error Messages + Postman Test Cases (#8)
* DataModel Initial Commit - ER Diagram, README, DDL Statements

* ignoring .code-workspace files (some IDE) and .DS_Store (mac users)

* Removing unwanted changes to IDE settings file + removed empty file

* Startup: Configure routing

I believe something to do with the jump to with the jump to .NET Core
3.1 required us to reconfigure this. This adds the middleware necessary
to route to the API controllers. It also adds a `TestController` as a
reference on how to add API endpoints.

* Remove useless files

We don't need any views since this is just a Web API. The
`DummyController` was just a demo and the corresponding table won't be
in our new database, so we should remove it. The sample data controller
should also be removed since we now have the `TestController` for
reference.

* Project: Remove unnecessary property groups

These are left over from when we had our front-end code in the same
project.

* API routing: Add default route attribute

* APIController attribute: Add to assembly

This gives us some nice perks discussed here:
https://www.strathweb.com/2019/01/enabling-apicontroller-globally-in-asp-net-core-2-2/

* removed weird spacing and drop [type] if exists methods

* Revert "removed weird spacing and drop [type] if exists methods"

This reverts commit 710eeb577b8bee9355b632b35dd0389e85aea2f0.

* removed spacing in graph added primary key to StoryCategory

* removed spaciing stuff from meta

* removed drop statements from Hourglass user

* removed geo map drop statements

* Added PublicationStatus to Story

* Removing .DS_Store

* Creating Skeleton Routes and Test Cases

Created a Controller for each entity in the database. Each controller is defined with four REST operations: GET, POST, PUT, DELETE. These controllers return "not implemented" messages when called. Each entity and rest operation contains a postman test associated with it. These tests are located in the `PostmanTests` folder under HourglassServer.

* Merging Master into SkeletonRoutesAndTestCases

* Moving Story Model Files into `Data` folder

* Adding Story + BlockContent Model files

Tested Server and confirmed Story is being parsed.

* Adding Database Entity Creator for Stories

* Updated Postman Tests with local environment

* Added DBContext to StoryController

* Renamed and Moved Story folder in `controllers` to `Data` folder

* Added Primary Key definition to Story

* Updated IDs to represent 36 length character UUIDs

* Removing request bodies from GET and DELETE (test cases too)

* Sample Story Code

This commit contains the sample code for creating, deleting, and retrieving (get) stories.

* Removing validity check for clarity sake

* Updating Data Model to be consistent with current database schema

* Removing DS Store and cleaning redundant gitignore line

* Removing empty default constructors

* Cleaning up empty/unused code

* Updated test cases to have lower case fields like our database

* Changing fields to be lowercase to match Postgres convention

* Removing unnecessary/ weird config action

* Updated Meta.sql

The copy you had wasn't the live version, so I updated it.

* Update Meta.sql

Co-authored-by: Zach Mitchell <zdmitche@calpoly.edu>
Co-authored-by: Justis Mackaoui <justismackaoui@gmail.com>