Adding script to generate classes from tables in database
This script is trying to tackle the problem that not every table in the database will receive a class on the backend. This script reads from HourglassServer/ablesToReverseEngineer.txt places them as arguments to the reverse engineering dotnet command, scaffold. Finally, this scaffold command is run and the classes generated.