Moving generated classes from `Data/Persistent` to `Model/Persistent` + updating HourglassContext
The convention is to output these files to the top-level `Models` folder. This folder did not exist previously and this commit it adding it, along with the models that were generated.

Also, HourglassContext was updated with the model information instead of having a whole new context (e.g. postgresContext) be generated.