Generate database models (#26)
* Adding script to generate classes from tables in database

This script is trying to tackle the problem that not every table in the database will receive a class on the backend. This script reads from HourglassServer/ablesToReverseEngineer.txt places them as arguments to the reverse engineering dotnet command, scaffold. Finally, this scaffold command is run and the classes generated.

* Adding instructions to README on how to run the script

Issue: https://stackoverflow.com/questions/44086572/cant-find-package-manager-console-in-visual-studio-for-mac
Summary: Powershell dependency not available to Mac or Linux users.
Resolution: Use .NET command line interface

* Adding new package for generated classes

Folder: /Data/Persistent

* Adding hardcoded list of tables to generate

This file hardcodes the names of the tables in order to avoid generating classes for user-uploaded tables, like datasets and graphs.

* Adding Generated Classes to Data/Persistent folder

* Adding instructions on how to specify which tables to reverse engineer

* ignore intellij files

* Adding `postgresContext` as dependancy to project through Startup.cs

* Updated README for formatting issues

1. Wrapped command in single quotes
2. Removed extra line

* Removing default graph table

Seems to be not updated and perhaps not being used in the database.

* Updating `postgresContext` to rename fix of table `Person`

* Refactored script to change lines in a file instead of substrings +  removed warning about connection string

Script automatically removed connection string so warning is no longer relevant.

* Replacing old Story persistence model with generated one.

* postgresContext formatting issues

Reran the script with the new refactored version.

* Updated script to use spaces instead of tabs to match other files in project

* new postgresContext with tabs replaces with spaces.

* Refactored script

- I separated the script's constituent parts into different files including: Constants + Removing the connection string.
- Models no longer being generated in `Data` folder (anti-convention)
- postgresContext no longer being generated. HourglassContext being updated instead.

* Updating `Person` to `person` in `TableToReverseEngineer.txt`

* Removing https route `api/education/degress `

This route is being used anymore. Furthermore, the entity `degrees` that it relied on is no longer a table in the database.

* Moving generated classes from `Data/Persistent` to `Model/Persistent` + updating HourglassContext

The convention is to output these files to the top-level `Models` folder. This folder did not exist previously and this commit it adding it, along with the models that were generated.

Also, HourglassContext was updated with the model information instead of having a whole new context (e.g. postgresContext) be generated.

* Updating references to `Person` to use generated class.

- Updating imports to use the `Models/Persistent` Person model
- Deleting old persistent model

* Updating `Person` + `HourglassContext` after pulling in database changes.

* Changing IEnumerable to ICollection

Co-authored-by: Alberto Rodriguez <vhsalbertorodriguez@gmail.com>