Person: implement put method
This will allow us to update a user's password. In the future we could
add functionality to update other aspects of a user as well. We need
this for password resetting.

In a future commit, we should add authentication to confirm that the
user is an admin or the user they're updating to avoid malicious
resetting of passwords among other things.