EmailService: extract to interface
Because this class acts as a service, it makes sense to abstract it to
an interface. Aside from being good practice, it can help us do things
like AB test different email bodies, subjects, etc. We are currently
experiencing a problem where our password email frequently ends up in
the user's spam folder. Now that this email service is abstracted, we
could easily write an AB test that gives the user the default email
service, or a new one that we haven't written yet. This new one could
format the email differently to see if a smaller percentage end up in
spam.

If in the future we wanted to create promotional emails or newsletter,
we could AB test to see which email service creates an email that leads
to more hits using a tool like Google Optimize.