Password reset: Send email (#62)
* Initial Email setup in PasswordController/PUT

* Utilities: add UpdateAsync

This fits right in with `InsertAsync` and `DeleteAsync`.

* Person: implement put method

This will allow us to update a user's password. In the future we could
add functionality to update other aspects of a user as well. We need
this for password resetting.

In a future commit, we should add authentication to confirm that the
user is an admin or the user they're updating to avoid malicious
resetting of passwords among other things.

* EducationController: remove file

This file is no longer needed.

* fixed return tag in PersonController

* PasswordController: cleanup

* PasswordController: make reset endpoint POST

* PasswordController: catch email errors

* Password reset: generate token to verfiy user

The token is added as a GET param to the url to reset the user's
password. This is for security reasons.

* PasswordController post: make async

* separate token functionality into login and passwordreset, edit PasswordController post to show that

* Utilities: add UpdateAsync

This fits right in with `InsertAsync` and `DeleteAsync`.

* Person: implement put method

This will allow us to update a user's password. In the future we could
add functionality to update other aspects of a user as well. We need
this for password resetting.

In a future commit, we should add authentication to confirm that the
user is an admin or the user they're updating to avoid malicious
resetting of passwords among other things.

* EducationController: remove file

This file is no longer needed.

* PasswordController: cleanup

* PasswordController: make reset endpoint POST

* PasswordController: catch email errors

* Password reset: generate token to verfiy user

The token is added as a GET param to the url to reset the user's
password. This is for security reasons.

* PasswordController post: make async

* separate token functionality into login and passwordreset, edit PasswordController post to show that

* Person: implement put method

This will allow us to update a user's password. In the future we could
add functionality to update other aspects of a user as well. We need
this for password resetting.

In a future commit, we should add authentication to confirm that the
user is an admin or the user they're updating to avoid malicious
resetting of passwords among other things.

* EducationController: remove file

This file is no longer needed.

* PasswordController: cleanup

* PasswordController: make reset endpoint POST

* PasswordController: catch email errors

* Password reset: generate token to verfiy user

The token is added as a GET param to the url to reset the user's
password. This is for security reasons.

* PasswordController post: make async

* separate token functionality into login and passwordreset, edit PasswordController post to show that

* Person: implement put method

This will allow us to update a user's password. In the future we could
add functionality to update other aspects of a user as well. We need
this for password resetting.

In a future commit, we should add authentication to confirm that the
user is an admin or the user they're updating to avoid malicious
resetting of passwords among other things.

* EducationController: remove file

This file is no longer needed.

* PasswordController: cleanup

* PasswordController: make reset endpoint POST

* PasswordController: catch email errors

* Password reset: generate token to verfiy user

The token is added as a GET param to the url to reset the user's
password. This is for security reasons.

* PasswordController post: make async

* separate token functionality into login and passwordreset, edit PasswordController post to show that

* Password reset: fix frombody error

* PasswordController POST: send email GET param

* PasswordController: run cleanup command

* PasswordController POST: send email GET param

* PasswordController: run cleanup command

* Password reset email: format body

Multiline strings need to be formatted against the left margin to avoid
extra spaces.

* Password reset role claim: make string constant

* Password PUT: cleanup

To avoid ambiguity with our routing system, we need to label the methods
Post and Put as HttpPost and HttpPut respectively. Also, let's move
email to the body of the put method rather than the URL. That seems to
be a more standard practice.

* added policy and handler for resetting password

* restructure token service

* SmtpClient: use MailMessage to send email

Using a class here is cleaner.

* Password reset token: move scope

This token should be generated in the scope it's used for clarity.

* EmailService: encapsulate email creation

Email creation is a separate responsibility than handling our Password
API requests. Let's move it into a new class.

* EmailService: extract to interface

Because this class acts as a service, it makes sense to abstract it to
an interface. Aside from being good practice, it can help us do things
like AB test different email bodies, subjects, etc. We are currently
experiencing a problem where our password email frequently ends up in
the user's spam folder. Now that this email service is abstracted, we
could easily write an AB test that gives the user the default email
service, or a new one that we haven't written yet. This new one could
format the email differently to see if a smaller percentage end up in
spam.

If in the future we wanted to create promotional emails or newsletter,
we could AB test to see which email service creates an email that leads
to more hits using a tool like Google Optimize.

* Cleanup: run Visual Studio cleanup command

* make Token/POST async, fix return object

* fix merge issue with new password hashing

Co-authored-by: Levi Thompson <leviwthompson@gmail.com>